+++ nuxt/app/components/nav/NavProjects.vue
