+++ nuxt/app/pages/auth/sign-in.vue
@@ -1,37 +1,27 @@
-<script lang="ts" setup>
-  import { signUp } from "~/utils/auth-client";
+<script setup lang="ts">
+  import { signIn } from "~/utils/auth-client";
   import { Eye, EyeOff } from "lucide-vue-next";
 
-  const firstName = ref("");
-  const lastName = ref("");
   const email = ref("");
   const password = ref("");
   const showPassword = ref(false);
   const isErrorDialogOpen = ref(false);
   const errorMessage = ref("");
 
-  const handleSignUp = async () => {
-    const user = {
-      firstName: firstName.value,
-      lastName: lastName.value,
-      email: email.value,
-      password: password.value,
-    };
-    await signUp.email({
-      email: user.email,
-      password: user.password,
-      name: `${user.firstName} ${user.lastName}`,
-      callbackURL: "/",
-      fetchOptions: {
+  const handleSignIn = async () => {
+    await signIn.email(
+      {
+        email: email.value,
+        password: password.value,
+        callbackURL: "/dashboard",
+      },
+      {
         onError(context) {
           errorMessage.value = context.error.message;
           isErrorDialogOpen.value = true;
         },
-        onSuccess() {
-          useRouter().push("/dashboard");
-        },
-      },
-    });
+      }
+    );
   };
 </script>
 
@@ -39,33 +29,29 @@
   <div class="flex h-screen items-center justify-center">
     <Card class="mx-auto max-w-sm">
       <CardHeader>
-        <CardTitle class="text-xl">Sign Up</CardTitle>
-        <CardDescription> Enter your information to create an account </CardDescription>
+        <CardTitle class="text-2xl"> Login </CardTitle>
+        <CardDescription> Enter your email below to login to your account </CardDescription>
       </CardHeader>
       <CardContent>
         <div class="grid gap-4">
-          <div class="grid grid-cols-2 gap-4">
-            <div class="grid gap-2">
-              <Label for="first-name">First name</Label>
-              <Input id="first-name" placeholder="Max" required v-model="firstName" />
-            </div>
-            <div class="grid gap-2">
-              <Label for="last-name">Last name</Label>
-              <Input id="last-name" placeholder="Robinson" required v-model="lastName" />
-            </div>
-          </div>
           <div class="grid gap-2">
             <Label for="email">Email</Label>
-            <Input id="email" type="email" placeholder="m@example.com" required v-model="email" />
+            <Input id="email" type="email" placeholder="m@example.com" v-model="email" required />
           </div>
           <div class="grid gap-2">
-            <Label for="password">Password</Label>
+            <div class="flex items-center">
+              <Label for="password">Password</Label>
+              <a href="/auth/forget-password" class="ml-auto inline-block text-sm underline">
+                Forgot your password?
+              </a>
+            </div>
             <div class="relative">
               <Input
                 id="password"
                 :type="showPassword ? 'text' : 'password'"
+                placeholder="password"
                 v-model="password"
-                placeholder="Enter your password"
+                required
               />
               <Button
                 type="button"
@@ -82,11 +68,11 @@
               </Button>
             </div>
           </div>
-          <Button type="button" class="w-full" @click="handleSignUp">Create an account</Button>
-        </div>
-        <div class="mt-4 text-center text-sm">
-          Already have an account?
-          <a href="/sign-in" class="underline"> Sign in </a>
+          <Button type="submit" class="w-full" @click="handleSignIn"> Login </Button>
+          <div class="mt-4 text-center text-sm">
+            Don't have an account?
+            <a href="/auth/sign-up" class="underline"> Sign up </a>
+          </div>
         </div>
       </CardContent>
     </Card>
