+++ nuxt/server/api/account/update-email.put.ts
