+++ nuxt/server/api/account/update-password.put.ts
